Move dummy implementation code to the testing module
